# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Layer para integrações de pagamento Afya.",
        "title": "Card Layer API",
        "contact": {
            "name": "Foundation Squad"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/customers": {
            "post": {
                "security": [
                    {
                        "x-api-key": []
                    }
                ],
                "description": "Cria um cliente e o tokeniza nos gateways de pagamento",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customers"
                ],
                "summary": "Cria um cliente",
                "parameters": [
                    {
                        "description": "Informações sobre o cliente",
                        "name": "customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payment-layer-card-api_usecases_customers_dtos.CreateCustomerDTOInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/payment-layer-card-api_usecases_customers_dtos.CreateCustomerDTOOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adapters_http.ErrorWrapperInternalDTO"
                        }
                    }
                }
            }
        },
        "/customers/document/{document}/serviceID/{serviceID}": {
            "get": {
                "security": [
                    {
                        "x-api-key": []
                    }
                ],
                "description": "Buscar um cliente pelo documento",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customers"
                ],
                "summary": "Obter um cliente pelo documento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Documento do cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/payment-layer-card-api_usecases_customers_dtos.CustomerOutputDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adapters_http.ErrorWrapperInternalDTO"
                        }
                    }
                }
            }
        },
        "/customers/{id}": {
            "get": {
                "security": [
                    {
                        "x-api-key": []
                    }
                ],
                "description": "Buscar um cliente pelo identificador",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customers"
                ],
                "summary": "Obter um cliente pelo id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identificador do cliente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/payment-layer-card-api_usecases_customers_dtos.CustomerWithTokensOutputDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/adapters_http.ErrorWrapperInternalDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "adapters_http.ErrorOutput": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "http_status": {
                    "type": "integer"
                },
                "log_message": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "adapters_http.ErrorWrapperInternalDTO": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/adapters_http.ErrorOutput"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "payment-layer-card-api_usecases_customers_dtos.Address": {
            "type": "object",
            "required": [
                "city",
                "country",
                "district",
                "line_1",
                "state"
            ],
            "properties": {
                "city": {
                    "description": "Cidade do endereço de cliente",
                    "type": "string",
                    "example": "São Paulo"
                },
                "country": {
                    "description": "País do endereço de cliente",
                    "type": "string",
                    "example": "BR"
                },
                "district": {
                    "description": "Bairro do endereço de cliente",
                    "type": "string",
                    "example": "Centro"
                },
                "line_1": {
                    "description": "Linha 1 do endereço de cliente",
                    "type": "string",
                    "example": "Rua João Silva, 123"
                },
                "line_2": {
                    "description": "Linha 2 do endereço de cliente",
                    "type": "string",
                    "example": "Apto 123"
                },
                "postalCode": {
                    "description": "CEP do endereço de cliente",
                    "type": "string",
                    "example": "12345678"
                },
                "postal_code": {
                    "description": "TODO: MUDAR PARA SNAKE CASE",
                    "type": "string",
                    "example": "12345678"
                },
                "state": {
                    "description": "Estado do endereço de cliente",
                    "type": "string",
                    "example": "SP"
                }
            }
        },
        "payment-layer-card-api_usecases_customers_dtos.CreateCustomerDTOInput": {
            "type": "object",
            "required": [
                "address",
                "birth_date",
                "document",
                "email",
                "first_name",
                "last_name",
                "phone",
                "service_id"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/payment-layer-card-api_usecases_customers_dtos.Address"
                },
                "birth_date": {
                    "type": "string",
                    "example": "2000-11-02"
                },
                "document": {
                    "type": "string",
                    "example": "12345678901"
                },
                "email": {
                    "type": "string",
                    "example": "teste@gmail.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "João"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female",
                        ""
                    ],
                    "example": "male || female"
                },
                "last_name": {
                    "type": "string",
                    "example": "Silva"
                },
                "phone": {
                    "type": "string",
                    "maxLength": 13,
                    "minLength": 10,
                    "example": "11999999999"
                },
                "service_id": {
                    "type": "string",
                    "example": "835524dd-d897-45ed-a68e-2e1a8bf522d7"
                }
            }
        },
        "payment-layer-card-api_usecases_customers_dtos.CreateCustomerDTOOutput": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Identificador do cliente",
                    "type": "string",
                    "example": "835524dd-d897-45ed-a68e-2e1a8bf522d7"
                }
            }
        },
        "payment-layer-card-api_usecases_customers_dtos.CustomerOutputDTO": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string",
                    "example": "2000-02-30"
                },
                "created_at": {
                    "type": "string",
                    "example": "2006-01-02T15:04:05-07:00"
                },
                "document": {
                    "type": "string",
                    "example": "12345678901"
                },
                "email": {
                    "type": "string",
                    "example": "teste@gmail.com"
                },
                "gender": {
                    "type": "string",
                    "example": "male || female"
                },
                "id": {
                    "type": "string",
                    "example": "835524dd-d897-45ed-a68e-2e1a8bf522d7"
                },
                "name": {
                    "type": "string",
                    "example": "João Silva"
                },
                "phone": {
                    "type": "string",
                    "example": "5511999999999"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2006-01-02T15:04:05-07:00"
                }
            }
        },
        "payment-layer-card-api_usecases_customers_dtos.CustomerTokens": {
            "type": "object",
            "properties": {
                "gateway": {
                    "type": "string",
                    "example": "pagarme"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                }
            }
        },
        "payment-layer-card-api_usecases_customers_dtos.CustomerWithTokensOutputDTO": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string",
                    "example": "2000-02-11"
                },
                "created_at": {
                    "type": "string",
                    "example": "2006-01-02T15:04:05-07:00"
                },
                "document": {
                    "type": "string",
                    "example": "12345678901"
                },
                "email": {
                    "type": "string",
                    "example": "teste@gmail.com"
                },
                "gender": {
                    "type": "string",
                    "example": "male || female"
                },
                "id": {
                    "type": "string",
                    "example": "835524dd-d897-45ed-a68e-2e1a8bf522d7"
                },
                "name": {
                    "type": "string",
                    "example": "João Silva"
                },
                "phone": {
                    "type": "string",
                    "example": "5511999999999"
                },
                "tokens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/payment-layer-card-api_usecases_customers_dtos.CustomerTokens"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "example": "2006-01-02T15:04:05-07:00"
                }
            }
        }
    },
    "securityDefinitions": {
        "x-api-key": {
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        }
    }
}